+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSB070SQQ3ZNRARZFHWHXPJ"
+        "01KSSN043J3RCG7P8R4ZN5NCE1"
     )
     r"""The unique identifier of the data source"""