+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSB070T3YH1QNPF50CM4GKG"
+        "01KSSN043K5W3JP4W0MGS7G58Q"
     )
     r"""The unique identifier of the data source"""