+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSB070YK7YE81MWNAAQMZ2H"
+        "01KSSN043PP6J6X6M385DKAC5Q"
     )
     r"""The unique identifier of the data source"""